playwright-python-suite/requirements.txt
playwright>=1.38.0
pytest>=7.0
pytest-asyncio
pillow
imagehash
requests
